Fixing list sentinel marker
I *knew* this was going to bite me. I wanted to have the bool in the
data union on each node so I didn't waste space, but that caused
pointers to be checked for truth values with caused an ephemeral bug
when detecting the sentinel. Easy fix is having the sentinel marker as a
separate bool on each node, though this wastes space. But if I want to
use a node as an iterator, I need to know whether or not it's the
sentinel, necessitating its marking.